create new component SoloProjectsCards.jsx to house mapData list and place it in the parent component SoloProjects
