+++ src/Home/Education.jsx
@@ -4,7 +4,7 @@ import PreviousExp from './PreviousExp'
 // DONE: create class for Box parent and Card child items
 // DONE: create show skills on hover of certificate
 // DONE: popup text overflows into next component when hover over CF cert
-// TODO: create move renderPreviousExp and renderRelatedEd into their own components, as children of education component
+// TODO: create move renderPreviousExp and renderRelatedEd into their own components PreviousExp and RelatedEducation, as children of education component
 
 const Education = () => {
   
@@ -12,7 +12,6 @@ const Education = () => {
     <>
       <RelatedEducation />
       <PreviousExp />
-
     </>
   );
 }
